Revert "Merge webpack-env.* into webpack.*"
This reverts commit b123aa5469db5e5a56f8517eca31c1cdaf569ed4.
